LayoutTests/editing/pasteboard/copy-paste-with-important-rules-expected.txt

This test ensures copying and paste respects !important in style rules.
To test manually, copy and paste the content in the first box to the second box. All text should remain in blue and should remain unboldened.

Original content:
| "
"
| <p>
|   "<#selection-anchor>hello "
|   <span>
|     style="color: red; font-weight: bold;"
|     "world"
| "
"
| <p>
|   <span>
|     class="Apple-style-span"
|     style="color: red; font-weight: bold;"
|     "WebKit<#selection-focus>"
| "
"

Pasted content:
| "
"
| <p>
|   "hello "
|   <span>
|     style="color: red; font-weight: bold;"
|     "world"
| "
"
| <p>
|   <span>
|     class="Apple-style-span"
|     style="color: red; font-weight: bold;"
|     "WebKit"
| "
"
